requirements.txt
pocketflow>=0.0.1
openai>=1.0.0
pyyaml>=6.0
python-dotenv~=1.1.1
sentence-transformers~=5.0.0
protobuf~=5.29.5
pytest~=8.4.1